Update Imprivata_User_Maintenance.ps1
Added a function to limit the script to a specific Imprivata SSO Policy.  This is good for users that have multiple policies or for instances that have a separate Confirm ID policy, where you would NOT want to remove users based on last logon date.  The new variable can be found on Line 6.